Node favor calculated in heartbeat
Instead of favor being statically given to a heartbeat message, favor is now calculated when the message is applied in heartbeat.py. Currently the favor is calculated only using latitude and longitude for distance based favor, but additional parameters can be added with (I imagine) no issue.

When running a node, command line arguments for the latitude and longitude can be added. Otherwise default to 0.0 (meaning that all nodes will have the same favor)